+++ src/OpenTelemetry/Internal/Builder/ProviderBuilderServiceCollectionExtensions.cs
@@ -19,6 +19,9 @@
 using System.Diagnostics;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using OpenTelemetry;
+using OpenTelemetry.Internal;
+using OpenTelemetry.Metrics;
 
 namespace Microsoft.Extensions.DependencyInjection;
 
@@ -37,6 +40,17 @@ internal static class ProviderBuilderServiceCollectionExtensions
         // those cases.
         services!.TryAddSingleton<IConfiguration>(sp => new ConfigurationBuilder().AddEnvironmentVariables().Build());
 
+        services!.RegisterOptionsFactory(configuration => new SdkOptions(configuration));
+
         return services!;
+    }
+
+    public static IServiceCollection AddOpenTelemetryMeterProviderBuilderServices(this IServiceCollection services)
+    {
+        services.AddOpenTelemetryProviderBuilderServices();
+
+        services.RegisterOptionsFactory(configuration => new MetricReaderOptions(configuration));
+
+        return services;
     }
 }